Debug, format date, vue package unique

+++ Models/DataBase/ITravelDAL.cs
@@ -6,8 +6,8 @@ namespace LittleBigTraveler.Models.DataBase
 {
     public interface ITravelDAL : IDisposable
     {
-        int CreateTravel(int destinationId, string departureLocation, DateTime departureDate, DateTime returnDate, double price, int numParticipants);
-        void ModifyTravel(int id, int destinationId, string departureLocation, DateTime departureDate, DateTime returnDate, double price, int numParticipants);
+        int CreateTravel(int destinationId, string departureLocation, DateTime departureDate, DateTime returnDate, int price, int numParticipants);
+        void ModifyTravel(int id, int destinationId, string departureLocation, DateTime departureDate, DateTime returnDate, int price, int numParticipants);
         void DeleteTravel(int id);
     }
 }
